LayoutTests/inspector/styles/metrics-box-sizing-expected.txt

Tests that content-box and border-box content area dimensions are handled property by the Metrics pane.

content-box
border-box
content-box rendered dimensions: 80 x 75
border-box rendered dimensions: 80 x 55

Running: testBorderBoxInit1

Running: testBorderBoxInit2

Running: testInitialBorderBoxMetrics
=== Initial border-box ===
margin: 1 1 1 1
border: 3 3 3 3
padding: 7 7 7 7
content: 35 x 35

Running: testModifiedBorderBoxMetrics
=== Modified border-box ===
margin: 1 1 1 1
border: 3 3 3 3
padding: 7 7 7 7
content: 60 x 35

Running: testContentBoxInit1

Running: testContentBoxInit2

Running: testInitialContentBoxMetrics
=== Initial content-box ===
margin: 1 1 1 1
border: 3 3 3 3
padding: 7 7 7 7
content: 55 x 55

Running: testModifiedContentBoxMetrics
=== Modified content-box ===
margin: 1 1 1 1
border: 3 3 3 3
padding: 7 7 7 7
content: 60 x 55
